Added modular surface render class
Created a surface render interface to be able to add more render targets.  This separated the framebuffer (godot Image class) object from the intermediate buffer so we can copy the frame from the core and then pass it to a render surface.  Now multiple target nodes can be targeted as a surface to display the video on.